+++ Core/Src/fft.c
@@ -6,7 +6,6 @@
  */
 
 #include "fft.h"
-#include "arm_math.h"
 
 arm_rfft_fast_instance_f32 S;
 float32_t hammingWindow[FFT_SIZE];
@@ -20,8 +19,8 @@ void FFT_Init() {
 	}
 }
 
-void FFT_Process(const uint16_t *adcBuffer, float *inputSignal,
-		float *fftOutputComplex, float *fftMagnitudes) {
+void FFT_Process(const uint16_t *adcBuffer, float32_t *inputSignal,
+		float32_t *fftOutputComplex, float32_t *fftMagnitudes) {
 	// Calcular y remover offset DC dinámico
 	float32_t mean = 0.0f;
 	for (int i = 0; i < FFT_SIZE; i++) {
@@ -42,7 +41,7 @@ void FFT_Process(const uint16_t *adcBuffer, float *inputSignal,
 	arm_cmplx_mag_f32(fftOutputComplex, fftMagnitudes, FFT_SIZE / 2);
 }
 
-float FFT_FindFundamentalFrequency(const float *fftMagnitudes) {
+float32_t FFT_FindFundamentalFrequency(const float32_t *fftMagnitudes) {
 //	int binMin = (int)(80.0f * FFT_SIZE / FS);
 //	int binMax = (int)(5000.0f * FFT_SIZE / FS);
 //
@@ -58,7 +57,7 @@ float FFT_FindFundamentalFrequency(const float *fftMagnitudes) {
 //
 //  return (maxIndex * FS) / FFT_SIZE;
 
-	float deltaF = FS / ADC_BUFFER_SIZE;
+	float32_t deltaF = FS / ADC_BUFFER_SIZE;
 
 	// Encontrar el pico máximo (ignorando DC)
 	int maxBin = 1;
@@ -70,11 +69,11 @@ float FFT_FindFundamentalFrequency(const float *fftMagnitudes) {
 		}
 	}
 
-	float fc = maxBin * deltaF;
+	float32_t fc = maxBin * deltaF;
 
 	// Buscar si hay energía en fc/2 (bin armónico inferior)
-	int subharmonicBin = (int) ((fc / 2) / deltaF + 0.5f);
-	float subharmonicVal = 0;
+	uint8_t subharmonicBin = (uint8_t) ((fc / 2) / deltaF + 0.5f);
+	float32_t subharmonicVal = 0;
 	if (subharmonicBin > 1 && subharmonicBin < ADC_BUFFER_SIZE / 2) {
 		subharmonicVal = fftMagnitudes[subharmonicBin];
 	}
@@ -87,12 +86,12 @@ float FFT_FindFundamentalFrequency(const float *fftMagnitudes) {
 	return fc;
 }
 
-float FFT_CalculateEnergy(float *fftOutput, int size) {
-	float totalEnergy = 0.0f;
+float32_t FFT_CalculateEnergy(float32_t *fftOutput, uint16_t size) {
+	float32_t totalEnergy = 0.0f;
 
 	// Calcular índice mínimo y máximo de interés en frecuencia
-	int binMin = (int) (80.0f * FFT_SIZE / FS);      // ~bin de 80 Hz
-	int binMax = (int) (5000.0f * FFT_SIZE / FS);    // ~bin de 5 kHz
+	uint16_t binMin = (uint16_t) (80.0f * FFT_SIZE / FS);      // ~bin de 80 Hz
+	uint16_t binMax = (uint16_t) (5000.0f * FFT_SIZE / FS);    // ~bin de 5 kHz
 
 	if (binMax > size)
 		binMax = size;
